only the creator can delete the post who have created post.

+++ routes/review.js
@@ -12,9 +12,11 @@ const isLoggedIn = require('../middlewares/isLoggedIn');
 router.post('/',isLoggedIn,async (req,res)=>{
     console.log('To recieve review for listing');
     let id = req.params.id;
-    let comment = req.body.comment;
-    let rating = Number(req.body.rating);
-    let nreview = new Review({comment,rating});
+    let nreview = new Review({
+        comment:req.body.comment,
+        rating:req.body.rating,
+        creator:req.user._id
+    });
     let card = await Property.findById(id)
     if(!card) throw new Error('Property not found');
     card.reviews.push(nreview._id);
@@ -28,6 +30,11 @@ router.post('/',isLoggedIn,async (req,res)=>{
 router.delete('/:idReview',isLoggedIn,async (req,res)=>{
     let pid = req.params.id;
     let rid = req.params.idReview;
+    let review = await Review.findById(rid);
+    if(review.creator.toString() !== req?.user._id.toString()){
+        req.flash('message', 'You are not authorized to delete this review');
+        return res.redirect(`/listing/${pid}`);
+    }
     await Property.findByIdAndUpdate(pid,{$pull:{reviews:rid}});
     await Review.findByIdAndDelete(rid);
     req.flash('message', 'Review Successfully Deleted');
